chore(server): core IoC #82 - (most) branches resolvers to TS
